Refactor blt auda to only run for sitenow sites
